editor/libeditor/base/tests/test_bug599983.html

<!doctype html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=599983
-->
<title>Test for Bug 599983</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<script src="/tests/SimpleTest/EventUtils.js"></script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=599983">Mozilla Bug 599983</a>
<div contenteditable>foo</div>
<script>
getSelection().selectAllChildren(document.querySelector("div"));
document.execCommand("bold");
is(document.querySelector("[_moz_dirty]"), null,
   "No _moz_dirty allowed in webpages");
</script>
